Change the respone of the basic_stat to show MOTD
According to the protocol definition, basic_stat contains MOTD.
The full_stat response comes with descriptions from the server
in a key, value format and what is MOTD in basic stat is returned
as __hostname__ in a full_stat by the server and is not defined
by this client.
And that is as in the protocol definition.
Closes #1