Fixed negative challenge tokens
The challenge token is an INT32 according to Dinnerbone ( https://dinnerbone.com/blog/2011/10/14/minecraft-19-has-rcon-and-query/ ), so it can have a minimum value of -2147483648 ( https://en.wikipedia.org/wiki/Integer_(computer_science) ) and not 0. I've ran in to problems while testing this module with the latest MCPE versions (Nukkit/Clearsky/Genisys), because they were all generating a negative challenge token (at least a negative challenge token was printed in the debug output) and the module threw an exception at line 65/66 in this file. I've added a new variable with the lower INT32 limit and changed the challenge token check to check against this variable.